SW: Hardware-testing fixes, QSPI icon store, and 48/64px fonts
- Auto-dimming, ESC/OK footer swap, resistance mode renames (Milliohmmeter/Ohmmeter)
- Fix voltmeter/60V readout stuck at zero (dead v_term_ext_mv_filt filter)
- Icon storage on external QSPI flash with load pipeline + flashIcons CLI command
- Fix UG_GetFontWidth/UG_GetFontHeight byte-offset bug in ugui.h
- Center passive-readout (Voltmeter/Ampmeter) big value in its window
- Settings/Display: show live backlight %, drop "not implemented" placeholder
- Add FONT_48X94/FONT_64X124 (DejaVu Sans Mono, Latin-1) via ttf2ugui, unused for now (flash budget)

+++ sw/bat_source/Core/Inc/adc.h
@@ -168,7 +168,9 @@ typedef struct {
 
     //uint16_t reference_poti;
 
-    uint16_t r_mOhmx10;
+    // uint32_t, not uint16_t: at 1mA excitation (Ohmmeter mode) resistances in
+    // the tens-of-ohms range already exceed a uint16_t's ~6.5 Ohm ceiling here.
+    uint32_t r_mOhmx10;
 
     uint16_t vref_mV;
